api/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
decorator==4.4.0
Django==2.2.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-guardian==1.5.1
djangorestframework==3.9.4
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.0.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
pluggy==0.11.0
pprint==0.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.0
pytest==4.5.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.2
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
